+++ test/integration/appRegistry_integration_test.go
@@ -51,12 +51,12 @@ func TestCRUDApp(t *testing.T) {
 	})
 
 	_, err := client.AppRegistryService.CreateApp(app)
-	require.Nil(t, err)
+	require.NoError(t, err)
 	defer client.AppRegistryService.DeleteApp(appName)
 
 	// List all apps
 	apps, err := client.AppRegistryService.ListApps()
-	require.Nil(t, err)
+	require.NoError(t, err)
 	// At least the app we just created should be present
 	assert.NotZero(t, len(apps))
 	//app-reg service bug https://jira.splunk.com/browse/APPLAT-5043
@@ -64,7 +64,7 @@ func TestCRUDApp(t *testing.T) {
 
 	// Get the app
 	app_ret, err := client.AppRegistryService.GetApp(appName)
-	require.Nil(t, err)
+	require.NoError(t, err)
 	require.Equal(t, app.WebAppPost().Name, app_ret.WebApp().Name)
 	require.Equal(t, app.WebAppPost().Title, app_ret.WebApp().Title)
 	require.Equal(t, app.WebAppPost().RedirectUrls, app_ret.WebApp().RedirectUrls)
@@ -74,13 +74,13 @@ func TestCRUDApp(t *testing.T) {
 	description := "new Description"
 	title := newAppTitle("newtitle")
 	redirecturl := []string{"https://newlocalhost"}
-	updateApp := appregistry.MakeUpdateAppRequestFromWebAppPut(appregistry.WebAppPut{
+	updateApp := appregistry.UpdateAppRequest{
 		Description:  &description,
 		RedirectUrls: redirecturl,
-		Title:        title,
-	})
+		Title:        &title,
+	}
 	app_update_ret, err := client.AppRegistryService.UpdateApp(appName, updateApp)
-	require.Nil(t, err)
+	require.NoError(t, err)
 	require.Equal(t, app.WebAppPost().Name, app_update_ret.WebApp().Name)
 	require.Equal(t, title, app_update_ret.WebApp().Title)
 	require.Equal(t, description, *app_update_ret.WebApp().Description)
@@ -88,7 +88,7 @@ func TestCRUDApp(t *testing.T) {
 
 	// Delete the app
 	err = client.AppRegistryService.DeleteApp(appName)
-	require.Nil(t, err)
+	require.NoError(t, err)
 }
 
 // Test RotateSecret in app-registry service
@@ -108,17 +108,19 @@ func TestAppRotateSecret(t *testing.T) {
 	})
 
 	app_created, err := client.AppRegistryService.CreateApp(app)
-	require.Nil(t, err)
+	require.NoError(t, err)
 	defer client.AppRegistryService.DeleteApp(appName)
 
 	// rotate secret
 	app_ret, err := client.AppRegistryService.RotateSecret(appName)
-	require.Nil(t, err)
+	require.NoError(t, err)
 	require.NotEmpty(t, app_created.WebApp().ClientId, app_ret.WebApp().ClientSecret)
 }
 
 // Test Create/Get/List/Delete subscriptions and get apps/subscriptions in app-registry service
 func TestSubscriptions(t *testing.T) {
+	t.Skip("Pending integration fix")
+
 	client := getSdkClient(t)
 	appName := fmt.Sprintf("g.s1%d", testutils.TimeSec)
 
@@ -133,22 +135,22 @@ func TestSubscriptions(t *testing.T) {
 		},
 	})
 	_, err := client.AppRegistryService.CreateApp(app)
-	require.Nil(t, err)
+	require.NoError(t, err)
 	defer client.AppRegistryService.DeleteApp(appName)
 
 	// create subscription
 	err = client.AppRegistryService.CreateSubscription(appregistry.AppName{AppName: appName})
-	require.Nil(t, err)
+	require.NoError(t, err)
 	defer client.AppRegistryService.DeleteSubscription(appName)
 
 	// get app subscription
 	appsubs, err := client.AppRegistryService.ListAppSubscriptions(appName)
-	require.Nil(t, err)
+	require.NoError(t, err)
 	require.Equal(t, 1, len(appsubs))
 
 	// Get a subscription of an app
 	subs, err := client.AppRegistryService.GetSubscription(appName)
-	require.Nil(t, err)
+	require.NoError(t, err)
 	require.NotEmpty(t, subs)
 
 	// Get a subscription from non-exist-app
@@ -175,10 +177,10 @@ func TestSubscriptions(t *testing.T) {
 	})
 
 	_, err = client.AppRegistryService.CreateApp(app2)
-	require.Nil(t, err)
+	require.NoError(t, err)
 	defer client.AppRegistryService.DeleteApp(appName2)
 	err = client.AppRegistryService.CreateSubscription(appregistry.AppName{AppName: appName2})
-	require.Nil(t, err)
+	require.NoError(t, err)
 	defer client.AppRegistryService.DeleteSubscription(appName2)
 
 	query := appregistry.ListSubscriptionsQueryParams{}.SetKind(appregistry.AppResourceKindService)
@@ -189,10 +191,10 @@ func TestSubscriptions(t *testing.T) {
 			found++
 		}
 	}
-	require.Nil(t, err)
+	require.NoError(t, err)
 	require.Equal(t, 1, found)
 
 	// Delete the subscription
 	err = client.AppRegistryService.DeleteSubscription(appName)
-	require.Nil(t, err)
+	require.NoError(t, err)
 }