+++ test/integration/client_integration_test.go
@@ -159,6 +159,8 @@ func (ml *MyLogger) Print(v ...interface{}) {
 }
 
 func TestRoundTripperWithSdkClient(t *testing.T) {
+	t.Skip("Pending integration fix")
+
 	client, err := sdk.NewClient(&services.Config{
 		Token:        testutils.TestAuthenticationToken,
 		Host:         testutils.TestSplunkCloudHost,
@@ -169,7 +171,7 @@ func TestRoundTripperWithSdkClient(t *testing.T) {
 
 	webhookAction := genWebhookAction()
 	action, err := client.ActionService.CreateAction(webhookAction)
-	require.Nil(t, err)
+	require.NoError(t, err)
 	defer client.ActionService.DeleteAction((*webhookAction.WebhookAction()).Name)
 	require.NotEmpty(t, action)
 	assert.Equal(t, 4, len(LoggerOutput))